# Benchmarks and performance improvements