--- a/Wangkanai.Benchmarks/Wangkanai.Benchmarks.csproj
+++ b/Wangkanai.Benchmarks/Wangkanai.Benchmarks.csproj
@@ -1,0 +1,16 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <OutputType>Exe</OutputType>
+        <TargetFramework>netcoreapp3.1</TargetFramework>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="BenchmarkDotNet" Version="0.12.1" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\src\Wangkanai.Detection.csproj" />
+    </ItemGroup>
+
+</Project>