--- a/Wangkanai.Benchmarks/Wangkanai.Benchmarks.csproj
+++ b/Wangkanai.Benchmarks/Wangkanai.Benchmarks.csproj
@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>netcoreapp3.1</TargetFramework>
+        <TargetFramework>netcoreapp3.0</TargetFramework>
     </PropertyGroup>
 
     <ItemGroup>