--- a/src/Models/UserAgent.cs
+++ b/src/Models/UserAgent.cs
@@ -8,15 +8,25 @@ namespace Wangkanai.Detection.Models
     public class UserAgent
     {
         private readonly string _useragent;
+        private readonly string _useragentLower;
 
         public UserAgent()
-            => _useragent = string.Empty;
+        {
+            _useragent      = string.Empty;
+            _useragentLower = string.Empty;
+        }
 
         public UserAgent(string useragent) : this()
-            => _useragent = useragent ?? string.Empty;
+        {
+            _useragent      = useragent ?? string.Empty;
+            _useragentLower = _useragent.ToLower();
+        }
 
         [DebuggerStepThrough]
         public override string ToString()
             => _useragent;
+
+        public string ToLower() 
+            => _useragentLower;
     }
 }