I've added benchmarks project and improved device detection hot path performance.

Core idea: 
- do not lowercase user agent string for every check (major perf boost)
- do not allocate delegates on array comparison (minor perf boost)

Hardware and Software

```
BenchmarkDotNet=v0.12.1, OS=Windows 10.0.17763.1158 (1809/October2018Update/Redstone5)
Intel Core i5-3570K CPU 3.40GHz (Ivy Bridge), 1 CPU, 4 logical and 4 physical cores
.NET Core SDK=3.1.102
  [Host]  : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
  LongRun : .NET Core 3.1.2 (CoreCLR 4.700.20.6602, CoreFX 4.700.20.6702), X64 RyuJIT
```

Before:

| Method |     Mean |    Error |   StdDev |   Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|--------:|------:|------:|----------:|
| Detect | 92.13 us | 0.397 us | 2.053 us | 44.9219 |     - |     - | 137.66 KB |

After: only `ToLower()`

| Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| Detect | 30.94 us | 0.195 us | 0.173 us | 7.6294 |     - |     - |  23.51 KB |

After: do not allocate delegates +  `ToLower()`

| Method |     Mean |    Error |   StdDev |   Median |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|---------:|-------:|------:|------:|----------:|
| Detect | 29.03 us | 0.094 us | 0.486 us | 28.95 us | 6.2256 |     - |     - |  19.13 KB |