@dv00d00 2020-04-22
Removed another large chunk of allocations. Tablet and Mobile collections were created every time they were accessed.

Current result is:

| Method |     Mean |    Error |   StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------- |---------:|---------:|---------:|-------:|------:|------:|----------:|
| Detect | 25.86 us | 0.497 us | 0.629 us | 4.4250 |     - |     - |  13.63 KB |